backend/requirements.txt
## The following requirements were added by pip freeze:
aiofiles==25.1.0
alembic==1.17.2
anyio==4.11.0
async-property==0.2.2
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
deprecation==2.1.0
ecdsa==0.19.1
exceptiongroup==1.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jwcrypto==1.5.6
jwt==1.4.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
python-dotenv==1.2.1
python-jose==3.5.0
python-keycloak==5.8.1
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
Flask-Login>=0.6.2

